refactor: move utils under search index module
Also move search index tests under
same test module.

Refs: LINK-2269
