+++ events/utils.py
@@ -3,11 +3,9 @@ import logging
 import re
 import warnings
 from datetime import datetime, timedelta
-from functools import cache
 from typing import Iterable, Optional
 
 import bleach
-import libvoikko
 import pytz
 from dateutil.parser import parse as dateutil_parse
 from django.conf import settings
@@ -24,10 +22,6 @@ from helevents.models import User
 
 logger = logging.getLogger(__name__)
 
-# setup libvoikko
-voikko = libvoikko.Voikko("fi", "/etc/voikko")
-voikko.setNoUglyHyphenation(True)
-
 
 def convert_to_camelcase(s):
     return "".join(word.title() if i else word for i, word in enumerate(s.split("_")))
@@ -37,71 +31,6 @@ def convert_from_camelcase(s):
     return re.sub(
         r"(^|[a-z])([A-Z])", lambda m: "_".join([i.lower() for i in m.groups() if i]), s
     )
-
-
-def get_field_attr(obj: object, field: str) -> str:
-    """
-    Get attr recursively by following foreign key relations.
-    :param obj: object to get the attribute from
-    :param field: field name or foreign key relation
-    :return: attribute value
-    """
-    fields = field.split("__")
-    if len(fields) == 1:
-        return getattr(obj, fields[0], None)
-    else:
-        first_field = fields[0]
-        remaining_fields = "__".join(fields[1:])
-        return get_field_attr(getattr(obj, first_field), remaining_fields)
-
-
-@cache
-def analyze_word(word: str) -> list:
-    return voikko.analyze(word)
-
-
-def get_word_bases(word: str) -> set:
-    """
-    Returns a list of word bases of the word.
-    :param word: the word to split
-    :return: a list of word bases
-    """
-    words = set()
-    word = word.strip()
-    analysis = analyze_word(word)
-    if len(analysis) == 0:
-        # if the word can't be analyzed, return the word itself
-        words.add(word)
-        return words
-    for item in analysis:
-        # extract the bases from the string
-        if "WORDBASES" in item:
-            word_bases = re.findall(r"\(([^+].*?)\)", item["WORDBASES"])
-            words.update(word_bases)
-        else:
-            words.add(word)
-    return words
-
-
-def split_word_bases(word: str, words: set, lang: str = "fi") -> set:
-    """
-    Splits the word into its bases and adds them to the set of words.
-    :param word: the word to split
-    :param words: the set of words to add the bases to
-    :param lang: the language of the word (default: "fi")
-    :return: the set of words
-    """
-    if lang == "fi":
-        w_array = word or []
-        if isinstance(word, str):
-            w_array = word.split()
-        for word in w_array:
-            word_bases = get_word_bases(word)
-            words.update(word_bases)
-    else:
-        # for other languages, just add the word itself
-        words.add(word)
-    return words
 
 
 def get_value_from_tuple_list(list_of_tuples, search_key, value_index):
